Count-down timers in matches (p1).
n.b. the "end game" mechanic is not yet completed, so the timer ultimately achieves nothing.

We now have a countdown timer for game modes. Much like Player and GUI enhancements, Gamemodes can be constructed via components. This means we can build new game modes super easily and really go wild.
This also means that we can create new game-modes within runtime, i.e. how in COD you can set the time limit, the map size, etc. This is super scalable, and will probably go on to be a big aspect of the "scale" of the game.