Better Damage Handling (precursor to range and world damage)
We need a way of differentiating not only last person to cause damage, but also last type of damage to be caused. For example, if we are hit with a ranged attack we may behave differently to a melee attack.
This functionality is now built in, and will be used as a precursor to ranged and world-based damage (fall damage, floor is lava, etc.), this is implemented after a quick GUI overhaul.