GUI health bar example & event dispatchers
I've added the Health Bar as a GUI element with notes so that we can (hopefully) follow on to make stamina check ups etc.

I've also converted a few functions into Event Dispatchers so that we can listen to them in the GUI, e.g. for displaying a gold star when we get added XP from a pickup.